// frontend/src/api/models/LocationAddRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LocationAddRequest
 */
export interface LocationAddRequest {
    /**
     * 
     * @type {number}
     * @memberof LocationAddRequest
     */
    x: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAddRequest
     */
    y: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAddRequest
     */
    z: number;
    /**
     * 
     * @type {string}
     * @memberof LocationAddRequest
     */
    name: string;
}

/**
 * Check if a given object implements the LocationAddRequest interface.
 */
export function instanceOfLocationAddRequest(value: object): value is LocationAddRequest {
    if (!('x' in value) || value['x'] === undefined) return false;
    if (!('y' in value) || value['y'] === undefined) return false;
    if (!('z' in value) || value['z'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function LocationAddRequestFromJSON(json: any): LocationAddRequest {
    return LocationAddRequestFromJSONTyped(json, false);
}

export function LocationAddRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationAddRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'x': json['x'],
        'y': json['y'],
        'z': json['z'],
        'name': json['name'],
    };
}

export function LocationAddRequestToJSON(json: any): LocationAddRequest {
    return LocationAddRequestToJSONTyped(json, false);
}

export function LocationAddRequestToJSONTyped(value?: LocationAddRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'x': value['x'],
        'y': value['y'],
        'z': value['z'],
        'name': value['name'],
    };
}