// frontend/src/api/models/LocationUpdateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LocationUpdateRequest
 */
export interface LocationUpdateRequest {
    /**
     * 
     * @type {number}
     * @memberof LocationUpdateRequest
     */
    x?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationUpdateRequest
     */
    y?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationUpdateRequest
     */
    z?: number;
    /**
     * 
     * @type {string}
     * @memberof LocationUpdateRequest
     */
    name?: string;
}

/**
 * Check if a given object implements the LocationUpdateRequest interface.
 */
export function instanceOfLocationUpdateRequest(value: object): value is LocationUpdateRequest {
    return true;
}

export function LocationUpdateRequestFromJSON(json: any): LocationUpdateRequest {
    return LocationUpdateRequestFromJSONTyped(json, false);
}

export function LocationUpdateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationUpdateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'x': json['x'] == null ? undefined : json['x'],
        'y': json['y'] == null ? undefined : json['y'],
        'z': json['z'] == null ? undefined : json['z'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function LocationUpdateRequestToJSON(json: any): LocationUpdateRequest {
    return LocationUpdateRequestToJSONTyped(json, false);
}

export function LocationUpdateRequestToJSONTyped(value?: LocationUpdateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'x': value['x'],
        'y': value['y'],
        'z': value['z'],
        'name': value['name'],
    };
}