// frontend/src/api/models/PersonResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Color } from './Color';
import {
    ColorFromJSON,
    ColorFromJSONTyped,
    ColorToJSON,
    ColorToJSONTyped,
} from './Color';
import type { Country } from './Country';
import {
    CountryFromJSON,
    CountryFromJSONTyped,
    CountryToJSON,
    CountryToJSONTyped,
} from './Country';
import type { LocationResponse } from './LocationResponse';
import {
    LocationResponseFromJSON,
    LocationResponseFromJSONTyped,
    LocationResponseToJSON,
    LocationResponseToJSONTyped,
} from './LocationResponse';

/**
 * 
 * @export
 * @interface PersonResponse
 */
export interface PersonResponse {
    /**
     * 
     * @type {number}
     * @memberof PersonResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof PersonResponse
     */
    name: string;
    /**
     * 
     * @type {Color}
     * @memberof PersonResponse
     */
    eyeColor?: Color;
    /**
     * 
     * @type {Color}
     * @memberof PersonResponse
     */
    hairColor: Color;
    /**
     * 
     * @type {LocationResponse}
     * @memberof PersonResponse
     */
    location?: LocationResponse;
    /**
     * 
     * @type {number}
     * @memberof PersonResponse
     */
    height: number;
    /**
     * 
     * @type {number}
     * @memberof PersonResponse
     */
    weight: number;
    /**
     * 
     * @type {Country}
     * @memberof PersonResponse
     */
    nationality?: Country;
}



/**
 * Check if a given object implements the PersonResponse interface.
 */
export function instanceOfPersonResponse(value: object): value is PersonResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('hairColor' in value) || value['hairColor'] === undefined) return false;
    if (!('height' in value) || value['height'] === undefined) return false;
    if (!('weight' in value) || value['weight'] === undefined) return false;
    return true;
}

export function PersonResponseFromJSON(json: any): PersonResponse {
    return PersonResponseFromJSONTyped(json, false);
}

export function PersonResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'eyeColor': json['eyeColor'] == null ? undefined : ColorFromJSON(json['eyeColor']),
        'hairColor': ColorFromJSON(json['hairColor']),
        'location': json['location'] == null ? undefined : LocationResponseFromJSON(json['location']),
        'height': json['height'],
        'weight': json['weight'],
        'nationality': json['nationality'] == null ? undefined : CountryFromJSON(json['nationality']),
    };
}

export function PersonResponseToJSON(json: any): PersonResponse {
    return PersonResponseToJSONTyped(json, false);
}

export function PersonResponseToJSONTyped(value?: PersonResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'eyeColor': ColorToJSON(value['eyeColor']),
        'hairColor': ColorToJSON(value['hairColor']),
        'location': LocationResponseToJSON(value['location']),
        'height': value['height'],
        'weight': value['weight'],
        'nationality': CountryToJSON(value['nationality']),
    };
}